# docs/04-api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NextCRM API",
    "description": "NextCRM is a comprehensive Customer Relationship Management system with multi-tenant architecture, subscription billing, and project management capabilities. This API provides programmatic access to all CRM features including accounts, contacts, leads, opportunities, invoices, projects, and document management.",
    "version": "3.0.0-alpha",
    "contact": {
      "name": "NextCRM Support",
      "url": "https://github.com/DrivenIdeaLab/nextcrm-app",
      "email": "support@nextcrm.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.nextcrm.io/api",
      "description": "Production API Server"
    },
    {
      "url": "https://demo.nextcrm.io/api",
      "description": "Demo API Server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/billing/create-checkout-session": {
      "post": {
        "summary": "Create Stripe checkout session",
        "description": "Create a Stripe checkout session for subscription upgrade",
        "tags": ["Billing"],
        "operationId": "createCheckoutSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["priceId"],
                "properties": {
                  "priceId": {
                    "type": "string",
                    "description": "Stripe Price ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/OwnerOnly"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from NextAuth authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error", "message"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or session invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthenticated",
              "message": "Valid authentication token is required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
...
      "OwnerOnly": {
        "description": "This endpoint is only accessible by organization owners",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Forbidden",
              "message": "Only organization owners can view billing subscriptions",
              "code": "OWNER_ONLY",
              "requiredRole": "OWNER"
            }
          }
        }
      },
...
    }
  }
}